+++ CHANGELOG.md
@@ -3,7 +3,7 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## 0.6.7 - 2019-##-##
-- DataFrame.fromPandas now uses index names as header names when available.
+- Add optional parameter index_names to DataFrame.fromPandas.
 
 ## 0.6.6 - 2019-05-11
 - Update internal library.